+++ packages/mui/src/components/ElementAvatar/getColor.ts
@@ -14,7 +14,15 @@
  * limitations under the License.
  */
 
-export * from './ElementAvatar';
-export * from './MuiCapabilitiesGuard';
-export * from './MuiThemeProvider';
-export * from './MuiWidgetApiProvider';
+export function getColor(id: string): string {
+  // Same colors and algorithm as Element is using, to get the same results:
+  // https://github.com/matrix-org/matrix-react-sdk/blob/667ec166d736dfb0ac49f67398a8b7a13db7d5ef/src/Avatar.ts#L91
+  const defaultColors = ['#0DBD8B', '#368bd6', '#ac3ba8'];
+  let total = 0;
+  for (let i = 0; i < id.length; ++i) {
+    total += id.charCodeAt(i);
+  }
+  const colorIndex = total % defaultColors.length;
+
+  return defaultColors[colorIndex];
+}